+++ nutype_macros/src/common/gen/mod.rs
@@ -180,7 +180,8 @@ pub trait GenerateNewtype {
         maybe_error_type_name: Option<ErrorTypeName>,
         traits: HashSet<Self::TypedTrait>,
         maybe_default_value: Option<syn::Expr>,
-    ) -> GeneratedTraits;
+        guard: &Guard<Self::Sanitizer, Self::Validator>,
+    ) -> Result<GeneratedTraits, syn::Error>;
 
     fn gen_new_with_validation(
         type_name: &TypeName,
@@ -283,7 +284,7 @@ pub trait GenerateNewtype {
             Self::TypedTrait,
             Guard<Self::Sanitizer, Self::Validator>,
         >,
-    ) -> TokenStream {
+    ) -> Result<TokenStream, syn::Error> {
         let GenerateParams {
             doc_attrs,
             traits,
@@ -328,9 +329,10 @@ pub trait GenerateNewtype {
             maybe_error_type_name,
             traits,
             maybe_default_value,
-        );
+            &guard,
+        )?;
 
-        quote!(
+        Ok(quote!(
             #[doc(hidden)]
             mod #module_name {
                 use super::*;
@@ -343,6 +345,6 @@ pub trait GenerateNewtype {
                 #implement_traits
             }
             #reimports
-        )
+        ))
     }
 }