+++ nutype_macros/src/integer/gen/traits/mod.rs
@@ -1,3 +1,5 @@
+mod arbitrary;
+
 use std::collections::HashSet;
 
 use proc_macro2::TokenStream;
@@ -14,18 +16,19 @@ use crate::{
         },
         models::{ErrorTypeName, TypeName},
     },
-    integer::models::{IntegerDeriveTrait, IntegerInnerType},
+    integer::models::{IntegerDeriveTrait, IntegerGuard, IntegerInnerType},
 };
 
 type IntegerGeneratableTrait = GeneratableTrait<IntegerTransparentTrait, IntegerIrregularTrait>;
 
-pub fn gen_traits(
+pub fn gen_traits<T: ToTokens>(
     type_name: &TypeName,
     inner_type: &IntegerInnerType,
     maybe_error_type_name: Option<ErrorTypeName>,
     traits: HashSet<IntegerDeriveTrait>,
     maybe_default_value: Option<syn::Expr>,
-) -> GeneratedTraits {
+    guard: &IntegerGuard<T>,
+) -> Result<GeneratedTraits, syn::Error> {
     let GeneratableTraits {
         transparent_traits,
         irregular_traits,
@@ -43,12 +46,13 @@ pub fn gen_traits(
         maybe_error_type_name,
         irregular_traits,
         maybe_default_value,
-    );
+        guard,
+    )?;
 
-    GeneratedTraits {
+    Ok(GeneratedTraits {
         derive_transparent_traits,
         implement_traits,
-    }
+    })
 }
 
 impl From<IntegerDeriveTrait> for IntegerGeneratableTrait {
@@ -114,6 +118,9 @@ impl From<IntegerDeriveTrait> for IntegerGeneratableTrait {
             IntegerDeriveTrait::SchemarsJsonSchema => {
                 IntegerGeneratableTrait::Transparent(IntegerTransparentTrait::SchemarsJsonSchema)
             }
+            IntegerDeriveTrait::ArbitraryArbitrary => {
+                IntegerGeneratableTrait::Irregular(IntegerIrregularTrait::ArbitraryArbitrary)
+            }
         }
     }
 }
@@ -147,6 +154,7 @@ enum IntegerIrregularTrait {
     Default,
     SerdeSerialize,
     SerdeDeserialize,
+    ArbitraryArbitrary,
 }
 
 impl ToTokens for IntegerTransparentTrait {
@@ -166,45 +174,51 @@ impl ToTokens for IntegerTransparentTrait {
     }
 }
 
-fn gen_implemented_traits(
+fn gen_implemented_traits<T: ToTokens>(
     type_name: &TypeName,
     inner_type: &IntegerInnerType,
     maybe_error_type_name: Option<ErrorTypeName>,
     impl_traits: Vec<IntegerIrregularTrait>,
     maybe_default_value: Option<syn::Expr>,
-) -> TokenStream {
+    guard: &IntegerGuard<T>,
+) -> Result<TokenStream, syn::Error> {
     impl_traits
         .iter()
         .map(|t| match t {
-            IntegerIrregularTrait::AsRef => gen_impl_trait_as_ref(type_name, inner_type),
-            IntegerIrregularTrait::Deref => gen_impl_trait_deref(type_name, inner_type),
+            IntegerIrregularTrait::AsRef => Ok(gen_impl_trait_as_ref(type_name, inner_type)),
+            IntegerIrregularTrait::Deref => Ok(gen_impl_trait_deref(type_name, inner_type)),
             IntegerIrregularTrait::FromStr => {
-                gen_impl_trait_from_str(type_name, inner_type, maybe_error_type_name.as_ref())
+                Ok(gen_impl_trait_from_str(type_name, inner_type, maybe_error_type_name.as_ref()))
             }
-            IntegerIrregularTrait::From => gen_impl_trait_from(type_name, inner_type),
-            IntegerIrregularTrait::Into => gen_impl_trait_into(type_name, inner_type),
+            IntegerIrregularTrait::From => Ok(gen_impl_trait_from(type_name, inner_type)),
+            IntegerIrregularTrait::Into => Ok(gen_impl_trait_into(type_name, inner_type)),
             IntegerIrregularTrait::TryFrom => {
-                gen_impl_trait_try_from(type_name, inner_type, maybe_error_type_name.as_ref())
+                Ok(gen_impl_trait_try_from(type_name, inner_type, maybe_error_type_name.as_ref()))
             }
-            IntegerIrregularTrait::Borrow => gen_impl_trait_borrow(type_name, inner_type),
-            IntegerIrregularTrait::Display => gen_impl_trait_dislpay(type_name),
+            IntegerIrregularTrait::Borrow => Ok(gen_impl_trait_borrow(type_name, inner_type)),
+            IntegerIrregularTrait::Display => Ok(gen_impl_trait_dislpay(type_name)),
             IntegerIrregularTrait::Default => {
                 match maybe_default_value {
                     Some(ref default_value) => {
                         let has_validation = maybe_error_type_name.is_some();
-                        gen_impl_trait_default(type_name, default_value, has_validation)
+                        Ok(gen_impl_trait_default(type_name, default_value, has_validation))
                     },
                     None => {
-                        panic!("Default trait is derived for type {type_name}, but `default = ` parameter is missing in #[nutype] macro");
+                        let span = proc_macro2::Span::call_site();
+                        let msg = format!("Trait `Default` is derived for type {type_name}, but `default = ` parameter is missing in #[nutype] macro");
+                        Err(syn::Error::new(span, msg))
                     }
                 }
             }
-            IntegerIrregularTrait::SerdeSerialize => gen_impl_trait_serde_serialize(type_name),
-            IntegerIrregularTrait::SerdeDeserialize => gen_impl_trait_serde_deserialize(
+            IntegerIrregularTrait::SerdeSerialize => Ok(gen_impl_trait_serde_serialize(type_name)),
+            IntegerIrregularTrait::SerdeDeserialize => Ok(gen_impl_trait_serde_deserialize(
                 type_name,
                 inner_type,
                 maybe_error_type_name.as_ref(),
-            ),
+            )),
+            IntegerIrregularTrait::ArbitraryArbitrary => {
+                arbitrary::gen_impl_trait_arbitrary(type_name, inner_type, guard)
+            }
         })
         .collect()
 }