+++ nutype_macros/src/string/gen/mod.rs
@@ -16,7 +16,10 @@ use crate::{
 
 use self::{error::gen_validation_error_type, traits::gen_traits};
 
-use super::{models::StringDeriveTrait, StringNewtype};
+use super::{
+    models::{StringDeriveTrait, StringGuard},
+    StringNewtype,
+};
 
 impl GenerateNewtype for StringNewtype {
     type Sanitizer = StringSanitizer;
@@ -172,12 +175,13 @@ impl GenerateNewtype for StringNewtype {
         maybe_error_type_name: Option<ErrorTypeName>,
         traits: HashSet<Self::TypedTrait>,
         maybe_default_value: Option<syn::Expr>,
-    ) -> GeneratedTraits {
-        gen_traits(
+        _guard: &StringGuard,
+    ) -> Result<GeneratedTraits, syn::Error> {
+        Ok(gen_traits(
             type_name,
             maybe_error_type_name,
             traits,
             maybe_default_value,
-        )
+        ))
     }
 }